move fuzzers
